Implemented view event for request authorization
